backend/requirements.txt
tensorflow
Pillow
numpy
Flask
flask-cors
matplotlib
opencv-python
tf-keras-vis
gunicorn